Move NativeCompiler into its own file.
