+++ lib/build.js
@@ -4,9 +4,9 @@ var fs = require("./file");
 var solc = require("solc");
 var constants = require("./constants");
 var fs = require("fs-extra");
+var NativeCompiler = require("./native_compiler");
 var os = require("os");
 var path = require("path");
-var child_process = require("child_process");
 var through = require('through2');
 var vinyl = require("vinyl-fs");
 var web3 = new require("web3")();
@@ -55,40 +55,3 @@ module.exports = class Builder {
         });
     }
 };
-
-class NativeCompiler {
-    static isAvailable() {
-        try {
-            child_process.execSync("solc");
-        } catch (err) {
-            return false;
-        }
-        return true;
-    }
-
-    static compile(input) {
-        var sources = input.sources;
-        var tmpDir = path.join(os.tmpdir(), 'dapple',
-                               String(Math.random()).slice(2));
-        fs.emptyDirSync(tmpDir); // Create or empty the directory. 
-
-        _.forIn(sources, (val, key) => {
-            fs.outputFileSync(path.join(tmpDir, key), val);
-        });
-
-        var output = JSON.parse(child_process.execSync(
-                "solc -o - --combined-json abi,bin,interface "
-                + Object.keys(sources).join(" "), {cwd: tmpDir}));
-
-        fs.removeSync(tmpDir); // Clean up.
-
-        output.contracts = _.mapValues(output.contracts, function (contract) {
-            return {
-                "bytecode": contract.bin,
-                "interface": contract.abi,
-                "solidity_interface": contract.interface
-            };
-        });
-        return output;
-    }
-}
